examples/base64_grayscale_service/pyproject.toml
[project]
name = "base64-grayscale-service"
version = "0.1.0"
description = "Example stateless service that converts base64 images to grayscale without S3."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "amplify-stateless>=1.0.0",
    "Pillow>=11.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["base64_grayscale_service"]

[tool.hatch.metadata]
allow-direct-references = true
